+++ intrinsics/fntransform.go
@@ -1,0 +1,8 @@
+package intrinsics
+
+func FnTransform(name string, input interface{}, template interface{}) interface{} {
+
+	// { "Fn::Transform" : { "Name" : "macro name", "Parameters" : {"key" : "value", ... } } }
+
+	return nil
+}