+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # GoFormation Versioning Changelog
 
+## [4.18.1](https://github.com/awslabs/goformation/compare/v4.18.0...v4.18.1) (2021-03-08)
+
+
+### Bug Fixes
+
+* **schema:** CloudFormation Updates ([#353](https://github.com/awslabs/goformation/issues/353)) ([eee9123](https://github.com/awslabs/goformation/commit/eee912305e08843fc112940333b93a6fc6aaf0ed))
+
+# [4.18.0](https://github.com/awslabs/goformation/compare/v4.17.0...v4.18.0) (2021-03-08)
+
+
+### Features
+
+* **intrinsics:** Add cloudformation.TransformFn() ([#352](https://github.com/awslabs/goformation/issues/352)) ([9a1e331](https://github.com/awslabs/goformation/commit/9a1e3315fd0dcbe2e9a1fcbd0994de7470a2bd9a))
+
 # [4.17.0](https://github.com/awslabs/goformation/compare/v4.16.4...v4.17.0) (2021-02-25)